add several more operators for Observable<Single>, Flowable<Single>, etc...
