+++ src/main/kotlin/io/reactivex/rxkotlin/maybe.kt
@@ -1,6 +1,8 @@
 package io.reactivex.rxkotlin
 
+import io.reactivex.Flowable
 import io.reactivex.Maybe
+import io.reactivex.Observable
 import java.util.concurrent.Callable
 import java.util.concurrent.Future
 
@@ -11,3 +13,38 @@ fun <T : Any> (() -> T).toMaybe(): Maybe<T> = Maybe.fromCallable(this)
 
 inline fun <reified R : Any> Maybe<Any>.cast(): Maybe<R> = cast(R::class.java)
 inline fun <reified R : Any> Maybe<Any>.ofType(): Maybe<R> = ofType(R::class.java)
+
+
+
+// EXTENSION FUNCTION OPERATORS
+
+/**
+ * Merges the emissions of a Observable<Maybe<T>>. Same as calling `flatMapMaybe { it }`.
+ */
+fun <T : Any> Observable<Maybe<T>>.mergeAllMaybes() = flatMapMaybe { it }
+
+/**
+ * Merges the emissions of a Flowable<Maybe<T>>. Same as calling `flatMap { it }`.
+ */
+fun <T : Any> Flowable<Maybe<T>>.mergeAllMaybes() = flatMapMaybe { it }
+
+/**
+ * Concatenates the emissions of an Flowable<Maybe<T>>.
+ */
+fun <T : Any> Observable<Maybe<T>>.concatAllMaybes() = concatMap { it.toObservable() }
+
+/**
+ * Concatenates the emissions of an Flowable<Maybe<T>>.
+ */
+fun <T : Any> Flowable<Maybe<T>>.concatAllMaybes() = concatMap { it.toFlowable() }
+
+/**
+ * Emits the latest `Maybe<T>` emitted through an `Flowable<Maybe<T>>`.
+ */
+fun <T : Any> Observable<Maybe<T>>.switchLatestMaybes() = switchMap { it.toObservable() }
+
+/**
+ * Emits the latest `Maybe<T>` emitted through an `Flowable<Maybe<T>>`.
+ */
+fun <T : Any> Flowable<Maybe<T>>.switchLatestMaybes() = switchMap { it.toFlowable() }
+
