+++ src/main/kotlin/io/reactivex/rxkotlin/subscribers.kt
@@ -1,10 +1,6 @@
 package io.reactivex.rxkotlin
 
-import io.reactivex.Completable
-import io.reactivex.Flowable
-import io.reactivex.Maybe
-import io.reactivex.Observable
-import io.reactivex.Single
+import io.reactivex.*
 import io.reactivex.disposables.Disposable
 import java.lang.RuntimeException
